refactor(tests): refactorizar módulo Products completo - dividir archivo grande
- Dividir products.service.spec.ts (1,088 líneas) en 4 archivos especializados:
  * products.service.crud.spec.ts (156 líneas) - Operaciones CRUD básicas
  * products.service.search.spec.ts (333 líneas) - Búsqueda y filtrado avanzado
  * products.service.updates.spec.ts (188 líneas) - Actualizaciones y cambios de estado
  * products.service.edge-cases.spec.ts (251 líneas) - Casos borde y validaciones
- Crear helper module products.test-helpers.ts (171 líneas) con:
  * createMockProduct() - Factory para entidades Product
  * createMockCreateProductDto() - Factory para DTOs de creación
  * createMockUpdateProductDto() - Factory para DTOs de actualización
  * createMockProductRepository() - Mock de repositorio
  * createMockQueryBuilder() - Mock de QueryBuilder
  * setupProductsTestModule() - Setup de módulo de test
  * setupQueryBuilderMocks() - Helper para configurar mocks de QB
- products.controller.spec.ts (271 líneas) - Sin cambios, ya dentro del objetivo
- Total: 81 tests pasando (100% pass rate)
- Coverage: Products module 96.42% statements, 88.23% branches
- Patrón aplicado: Factory pattern, DRY, helpers compartidos
- Files: 1,088 → 1,370 líneas (5 archivos + helpers)
